Refactored the structure of Animation and commented out Fluid due to previous structure overhaul
